Add CallWithNewTarget
Summary:
CallWithNewTarget is a generalized version of Call and Construct. The
latter implicitly initialize new.target to either undefined or the
callee, while the new instruction allows it to be explicitly specified.
It is theoretically possible to eliminate Call and Construct and replace
them with CallWithNewTarget, but it would, of course, be a code size
regression because of the extra argument. This new bytecode
instruction is not emitted anywhere yet.

Reviewed By: avp

Differential Revision: D52094773

fbshipit-source-id: 8b8e921beb2203f77c14e3ca7f3b16448723130d