Fix HBC backend for function calls
Summary:
Add support for `new` in the HBC backend, which was broken when we
removed ConstructInst. Also use CallWithNewTargetLong for all cases of
number call arguments > 256. This requires a new pass before register
allocation (which we merge with `FuncCallNOpts`) that will lower
CallInst into the appropriate instruction. Also, modify
`ISel::generateCallInst` to correctly handle all the potential
instructions.

Reviewed By: avp

Differential Revision: D51952271

fbshipit-source-id: c34f5c98435826741e83c5e89fbec434bbe7efc5