Remove handle from StringPrimitive::createEfficient()
Summary:
`createEfficient()` was unnecessarily allocating a Handle. In turn, one
specific caller - `_sh_asciiz_to_string()` - was not creating a marker,
since it did not expect it was necessary. The result was a huge
allocation of handles and a 2x performance regression in one test.

Remove the handle and add NoHandleScope in `_sh_asciiz_to_string()`.

Reviewed By: neildhar

Differential Revision: D52649629

fbshipit-source-id: 5b35a3bf5324b82bd1d3aedd96d4e985e6e540dd