Fix TypeInference by expanding param and return types
Summary:
Properly expand param and return types from NoType instead of defaulting
them to 'any'.

This was causing infinite loops in simple test cases like
```
function f() {
  return f();
}
```
so we fix that in this diff.

We were returning `changed` when nothing had changed to indicate there
was an instruction that we wanted to return to, but we should only
return when the operands types have actually changed.

Account for unreachable/non-returning code by resetting NoTypes to their
prepass types at the end of the loop.
In the following diffs, we can set attributes to indicate reachability
and whether functions return, and use that information to delete the
unused code.

Reviewed By: neildhar

Differential Revision: D52635993

fbshipit-source-id: 35ff89eaea4459b495dd3d6eb8496445bc78cf9c