Update codegen to output type guards instead of `%checks`
Summary:
Original Author: samzhou19815@meta.com
Original Git: 4c7e6d4f32cdc7fd382c7444660835d10f37bc95
Original Reviewed By: alexmckenley
Original Revision: D51512414

This diff updates the hermes-parser codegen to output type guards instead of `%checks` for the generated predicates file.

The current setup is in a weird state (we can't use hermes-parser to parse new syntax, because we are in hermes-parser itself), so I have to use flow-comments and disable prettier in the generated predicates file (prettier is on 2.8.8, which uses an old bundled-in flow-parser that doesn't support the new syntax). This will be a blocker for removing the remaining `%checks` or for adopting any new flow syntax in the hermes-parser codebase. At some point, we should just deal with the pain and move everything to prettier 3

Reviewed By: tmikov

Differential Revision: D52854404

fbshipit-source-id: 508da8ad88388beb6edd344e2bc9944a7326cad4