Implement HostObject
Summary:
Original Author: neildhar@meta.com
Original Git: cb26bbe5dcb995122a9da102b0b6365fbf7e79dc
Original Reviewed By: avp
Original Revision: D51551526

Implement HostObject in the ABI and JSI. This includes implementing a
new `HermesABIPropNameIDList` to pass the result of
`HostObject::getPropertyNames`.

The ABI method for getting properties from a `HostObject` is named
slightly differently than its JSI counterpart (`get_own_keys` vs
`getPropertyNames`). This is to align it with `Reflect.ownKeys` and the
Proxy `ownKeys` trap, and permit `Symbol`s to be returned from the
function. Our implementation doesn't currently deal with such symbols
properly, but that is a separate concern.

JSI is currently ambiguous on whether this is permissible, but this is
something we should address and align the two on before we stabilise
the ABI.

Reviewed By: tmikov

Differential Revision: D52854431

fbshipit-source-id: 1edd229fc95b9450cf46f37d98b052ca2ed58b4b