Update the Github issue template
Summary:
Original Author: tmikov@meta.com
Original Git: 4064e8709203f3b4506c76e32716eeccba5fbac1
Original Reviewed By: neildhar
Original Revision: D52580926

The issue template was misleading by implying that we might be able to
support non-current versions of Hermes, which is definitely not the
case. React Native occasionally backports fixes, but that is only after
the fix has been applied to the current version.

The stable ABI will improve this, but we are not there yet.

I also attempted to address a couple of things that bugged me:
- Asking for a Hermes version is nonsensical and has been for a long
time.
- It made it seem like reporting the OS is optional.

Frankly, I am still not very happy with the template, but I hope that at
least these are changes in the right direction.

Reviewed By: tmikov

Differential Revision: D52854405

fbshipit-source-id: a441d592d8b99996c46189bb8bd01bdba76dbda9