Support component syntax lowering
Summary: Add the component syntax lowering pass. Note this wont work for SH as is, as the outputted types are incorrect (the transform outputs a reference to `React.Node` but since the bundler inlines references it should be `react_index$Node`. This however is fine for non SH environments where we will strip the types anyway.

Reviewed By: josephsavona

Differential Revision: D52986713

fbshipit-source-id: a8ff10372872f00e99d135ef479ab931d14d782e