Add generic info storage to FlowChecker
Summary:
FlowChecker needs to track information about generics so that it can
instantiate them.

Put the relevant information into a `GenericInfo` struct and store it in
`FlowChecker` for the duration of its run.
There's no need to persist this information after the typechecker
finishes because IRGen doesn't know/care about generics because the AST
will be duplicated and the semantic data will be updated to point to
specializations.

We unique the specializations based on the list of `TypeInfo *` used
as the type arguments to the generic, so that we can correctly
deduplicate specializations in the following case:
```
function foo<T>() {}
foo<number | string>();
foo<number | string>(); // same specialization
```
where the two calls have different `Type *` due to different AST nodes
introducing the two unions.
The two `TypeInfo *` will be equal when `TypeInfo::equals()` is called.

Reviewed By: tmikov

Differential Revision: D49884368

fbshipit-source-id: 39462e82e406337a33690da268f73ef22e344f3b