Register and specialize generic functions in FlowChecker
Summary:
FlowChecker can find generic functions and record them,
along with any state needed to specialize them.

When encountering a callsite, the AST is copied, the new function is
typechecked, and the callsite is modified to point to the new function.

We prohibit uses of the generic functions outside of calls right now,
because the functions aren't real outside of calls.

The next steps will be generic classes.

Reviewed By: tmikov

Differential Revision: D49884367

fbshipit-source-id: ab084e67e78e5cf0116f6c8a2a7f96ff8c931d7f