Simple type refinement via checked casts
Summary:
A type refinement implementation which assumes that the Flow typechecker
has run prior to SH invocation (and the execution of FlowChecker.cpp).

When the typechecker encounters assignment, call, or member expression,
involving a union-typed expression, it attempts to narrow the
expression's type to a single non-optional type, inserts an implicit
checked cast, and does not emit an error.

This is a conservative set of AST nodes which allow this new refinement,
but it allows for a much more user-friendly way of dealing with
nullable/optional values.

More positions can be added in the future, but for now, this is a simple
subset of features to add support for the casts to.

Reviewed By: neildhar

Differential Revision: D53333545

fbshipit-source-id: 1b42b1e77081522feb3204b756f593b44bd08ded