Implement generic classes
Summary:
Class generics work somewhat similarly to function generics.

However, because we need to parse class types and populate them in the
binding table, we need to perform some extra steps.

## Defer class parsing during DeclareScopeTypes

During DeclareScopeTypes, we don't yet know all the classes that can be
referenced from within each of the class bodies.
This means we don't know if we're currently parsing them,
so we can't recursively parse newly specialized generic classes.
Example:
```
class A<T> {
  b: B<T>;
}
class B<T> extends A<T> {}
new A<number>();
```
If we parsed all the types eagerly, we would fail because we would
attempt to parse `A<number>` -> `B<number>` -> `A<number>`.

So we defer parsing during DeclareScopeTypes and collect all the
deferred generic class specializations and parse them all at once
at the end of DeclareScopeTypes.

It is possible that there are _new_ generic class specializations that
end up deferred during the parsing phase, so they will be handled during
that time as well.

## Enqueue typechecks of specialization bodies

After parsing, we have the question of when to descend into the class
bodies and typecheck the methods.

When we are running the typechecker on some expression and we create a
generic specialization, we can't immediately typecheck the resultant
class specialization method bodies, but we need the type so that we can
immediately typecheck the methods at the callsite.
Example:
```
class A<T> extends B<T> {
  constructor() {
    super();
  }
  foo(): void {}
}

class B<T> {
  constructor() {}
  method(): void {
    new A<boolean>().foo();
  }
}

new A<number>();
```
In this case, if we were to eagerly typecheck class method bodies,
we would typecheck:
```
A<number>
B<number> (via superClass)
A<boolean> (via NewExpression)
B<boolean> (via superClass)
A<boolean> <- ERROR: A<boolean> not initialized yet
```
and `A<boolean>` wouldn't be initialized yet.

Therefore, we queue all the generic specialization body typechecks until
after we have run the rest of the typechecker.
This is a simple way of preventing introducing new types that weren't
discovered during DeclareScopeTypes during the course of regular
typecheck visitation.

Currently this is the only use case that the typecheck queue supports,
but it's easy to add other cases if necessary.

Reviewed By: tmikov

Differential Revision: D51328087

fbshipit-source-id: 6a11c649565215fdb2988f0fe0fe4b360e3a4708