Rust crates for estree representation/parsing/analysis (#1276) (#1276)
Summary:
Original Author: joesavona@fb.com
Original Git: 6f16d42f9dae6522c087cd355f36df5e794585a6
Original Reviewed By: tyao1
Original Revision: D53071562

(discussed internally with pieterv but sharing for broader context)

This PR adds several crates designed to expose HermesParser to Rust to support parsing JS source code into an ergonomic Rust representation of the AST, serializing that format to/from estree json, and performing semantic analysis of the AST. The crates are structured as follows:

* hermes_estree is an ergonomic Rust representation of a JavaScript AST using a structure that aligns as closely as possible to the estree spec, and which is designed for accurate serialization to/from estree-compatible JSON.
* hermes_estree_codegen is the code generator for the above crate, driven from an easy to edit JSON-based description of the AST.
* hermes_parser wraps the lower-level hermes parser crate, and converts raw parsed nodes into the above estree format.
* hermes_semantic_analysis does name resolution for identifiers and labels, and checks other properties such as syntactic breaks/continue.
* hermes_diagnostic is used for representing invalid syntax, todos, compiler invariants, etc.
* hermes_utils is for utils.

Note that the only crate here which specifically depends on Hermes code is `hermes_parser`, which bridges the raw Hermes parser result into the estree format. The estree format is Hermes-agnostic, the semantic analysis is written against estree rather than the raw hermes parse tree, etc.

So we could put most or all of this code somewhere else, but it isn't clear where.

Our motivation for landing this code now is that we'd like to use this within Relay Compiler. In the future, it may also be used for the upcoming React Compiler (which was the original motivation for writing this code).

Pull Request resolved: https://github.com/facebook/hermes/pull/1276

Pulled By: josephsavona

Reviewed By: avp

Differential Revision: D54092055

fbshipit-source-id: 7803adab0e9c7713babe18584657baf5758f08f5