Compute scope depth for resolution during lowering
Summary:
Instead of doing the scope depth computation for scope resolution in
isel, do it during lowering and record the depth in the instruction.
This brings the lowered instruction closer to what is actually emitted,
and avoids doing anything complicated in isel.

More importantly, this keeps the structure of the IR intact during
scope resolution, since no Movs have been inserted. This is a useful
property in subsequent diffs.

Reviewed By: avp

Differential Revision: D53946804

fbshipit-source-id: 3146fd822d7309fd16e8d07e4ef52360902eb5ac