Fix equality of NaNs in isSameValue
Summary:
Original Author: neildhar@meta.com
Original Reviewed By: tmikov
Original Revision: D54179136

`isSameValue` and `isSameValueZero` need to return true if both the
arguments are `NaN`. Our current implementation does this by falling
back to a bitwise comparison in the case that both arguments are
numbers.

However, this is insufficient. There are two possible encodings for the
quiet NaN (with and without the sign bit), and it is easy to end up
with a mix of them in a single program. For example, the constant
`std::numeric_limits<double>::quiet_NaN` sets the sign bit to 0,
whereas `hermesc` currently always sets it to 1. On x86, operations
that produce NaN will set the sign bit to 1 if neither operand is a
NaN, whereas on ARM the default seems to be 0.

In principle, we could use that information to impose a canonical
encoding at runtime (using 1 on x86 and 0 on ARM), with checks during
deserialisation from bytecode to make sure we get the correct one for
the platform. But this would likely be fragile and complicated.

This diff patches `isSameValue` to check explicitly for NaN with
`std::isnan`, and make sure it always compares equal.

Note that this issue is further complicated in Hermes today because the
interpreter decodes the serialised NaN with
`encodeUntrustedNumberValue`, which means that the sign bit added by
`hermesc` is removed at runtime. This means that the added test case
only fails on x86, since on ARM, all NaNs end up without the sign bit.

Reviewed By: avp

Differential Revision: D54293380

fbshipit-source-id: ba0557f1a0ddf693a9a66bfc3c444b6f140b1140