Track dominance in FunctionAnalysis
Summary:
When forwarding the closure or its operands to a user, we need to
ensure that the closure is not only in the same function, but
actually dominates the user. In principle, it is possible to go through
the frame to a user that is not actually statically dominated, for
instance.

This diff adds a conservative mechanism for tracking dominance, by
keeping a bit that records whether we have passed through the frame.

Reviewed By: avp

Differential Revision: D54381109

fbshipit-source-id: 9cfc9012e6ace12c9b1a70810805d0518ffa4da2