Optimise ResolveScope
Summary:
Optimise `ResolveScope` in `InstSimplify`, by trying to walk its
starting scope parameter until we either get to the target scope, or
move up the chain to a scope that is closer to the target. This makes
`ResolveScope` more efficient, and allows any unnecessary
`CreateScopeInst`s along the way to be eliminated by DCE, since the
`ResolveScope` will no longer be using them.

Reviewed By: avp

Differential Revision: D53462040

fbshipit-source-id: 0c6e19c1a31cbec4e7eaa4ab5f7d209df892f45c