Remove owned variable analysis in FrameLoadStoreOpts
Summary:
Remove the entry block only analysis in FrameLoadStoreOpts, in order to
simplify preparing it for block scoping. This analysis can still be
done, but it should likely be a separate function or even a separate
pass to factor the complexity involved. When we do that, we can also
extend the analysis past the entry block.

Reviewed By: avp

Differential Revision: D54541590

fbshipit-source-id: b97444c0008f8d085774a219efa71848628c42a4