Use scope creation for TDZ checks in IRGen
Summary:
Now that scopes are explicit, instead of looking at the parent of a
VariableScope to determine if it can be analysed for TDZ violations in
IRGen, instead look at whether the operand of the read/write is a
CreateScopeInst.

In addition, instead of just recording variables that have been
initialised, also record the associated CreateScopeInst. This is not
strictly necessary in IRGen, since we can currently only have a single
CreateScopeInst for a given VariableScope. But given that the IR itself
will impose no such restriction, this seems like the right thing to do.

Reviewed By: avp

Differential Revision: D54545170

fbshipit-source-id: 87383b9f324d1fc1e3c3888dd2a4e3e2b312b1fb