Enqueue the implicit ctor for compilation, rather than emitting directly. Necessary for field inits.
Summary:
This diff makes the code generation for an implicit constructor obey
the discipline of enqueuing a compilation function, rather than doing
the code generation "directly."  This is necessary for a later diff in
the stack: we'll need to ensure that the field initializer function
gets compiled before the implicit ctor, since the latter may call the
former.

(Actually, I now wonder if that's true -- maybe all that needs to be
true is that variable holding the field initializer function is
available, and the order of compilation doesn't matter.  Still, it
seems like a good idea to have a consistent idiom for all
compilations.)

This also introduces some changes for "implicit functions" that are
also used for the field init function later:

* The FunctionInfo semantic info for a function may not have an AST
FunctionLikeNode; for implicit function associated with a class we
will pass the ClassDeclarationNode.

* In other cases we will pass null as the Function, and some functions
must be adjusted to handle this situation correctly; for example
counting arguments.

Reviewed By: avp

Differential Revision: D54313169

fbshipit-source-id: d146fe2c69187383c38b07633b29dfac35e1b70e