Clean up BytecodeProviderFromSrc creation methods
Summary:
There were too many factory methods to keep track of here,
and they all just passed through to a single function anyway.

Remove most of the functions, and only keep two ways to make a
`BCProviderFromSrc`: by compiling JS, or from an existing
`hbc::BytecodeModule`, the latter of which is just a simple passthrough
to the constructor.

Reviewed By: neildhar

Differential Revision: D54876669

fbshipit-source-id: 34c37ecb09f050a9a3090d74c655ac0f0db9687d