Fix IRVerifier calls in opt pipeline
Summary:
Fix a double-print issue when an error occurs during the optimization
pipeline. We should not be calling `verifyModule` after running the
optimizations anymore, since the PassManager will run it in-between each
run if the IRVerfifier is enabled.

However, we should be verifying the IR right after IRGen to check for
any errors that were not due to the opt pipeline.

Reviewed By: avp

Differential Revision: D54824213

fbshipit-source-id: 62f2c25166eb94b3851bf7ef0db342d18aea1e3c