Move UniquingRegExpTable into BytecodeModule
Summary:
`UniquingRegExpTable` can be kept around in `BytecodeModule`
for scenarios where we may want to modify it later.

It's possible that we can eventually shrink the amount of data that it
actually stores to save some memory, but for now just move it.

We need to maintain the mapping and the storage needs to be accessible
to the `UniquingRegExpTable`, so make it eagerly and store it.
This makes`getEntryList` and `getBytecodeBuffer` simple inline
functions.

`UniquingRegExpTable` stores pointers to `CompiledRegExp`, which lives
in the `Context`, so now we need to ensure the `Context` outlives
`BytecodeModule` (which in non-lazy mode it currently does, but perhaps
we will also store a `shared_ptr<Context>` in `BytecodeModule` in the
future).

Reviewed By: neildhar

Differential Revision: D55205911

fbshipit-source-id: 563b50f6ac165e2b4233b9e9fae2ac70fec824bb