Ensure IRVerifier labels are consistent with the dumped output
Summary:
Sometimes the label numbers generated in the IRVerifier instruction
messages don't line up with the numbers in the dump. After doing some
digging, it turns out that the namer holds this BB/instruction label on
a per-function basis. Therefore, before generating the label in
IRVerifier, we must restore this state by calling newFunction with the
correct function.

Reviewed By: avp

Differential Revision: D55215219

fbshipit-source-id: 0525c92e1ba3c132c0d0b3c802d73a97e0abb214