Simple Mem2Reg
Summary:
Implement a new SSA conversion pass based on the simple SSA
construction paper.

This produces slightly fewer Phis in existing code, since it does not
emit Phis for dead stores (see the changes in `unused-phi-cycle.js`).

The pass takes the following steps:
1. Prepare the IR such that each block only has at most a single load
and store for a given stack location. If both are present, the load
must precede the store. This means that any load must resolve to a
value in a predecessor block, and any store must be the live-out value
from its block for that stack location.
2. Replace every load with a placeholder Phi placed at the start of its
block.
3. Populate every inserted Phi with the live-out values of its
predecessors. If a block does not already have a live-out value, insert
a placeholder Phi in that block to serve that purpose.
4. Delete the (many) redundant Phis inserted by the previous steps.
5. Delete store-only stack locations.

Reviewed By: tmikov

Differential Revision: D45997889

fbshipit-source-id: a3303f5ee8726cad4c78a48e97e2e3407d39c504