Delete SH-specific LowerAllocObjectLiteral pass
Summary:
The `LowerAllocObjectLiteral` pass in the common Lowering file now
commits objects with at least 1 property to the buffer. So there is no
need to maintain a separate pass in SH anymore. The only motivating
factor previously was to commit smaller objects to the buffer than were
being allowed in the common lowering pass.

Reviewed By: neildhar

Differential Revision: D55621360

fbshipit-source-id: 3db3769a4043a5f5cc09f0eda98ea0919a8cc966