Dump scopes separately from functions
Summary:
Once VariableScope and Function are decoupled, it will no longer make
sense to dump the scope alongside a function. This diff changes
VariableScopes dumping so they are instead dumped right before the
first function that uses them.

Scopes are now dumped in the following form:
```
scope %VS1 [foo: any, bar: number]
```

This diff also changes the dumping format for Variables, such that
they are now dumped in the form `[%VS0.foo]`, where `%VS0` is the
VariableScope owning that Variable, and `foo` is the name of the Variable.

Reviewed By: avp

Differential Revision: D54545173

fbshipit-source-id: eedfc0de3283c8c5b215e44265dbddd438bfbbce