Relax requirements on scope operands
Summary:
Eliminate the requirement that all scope operands must be
`BaseScopeInst`. This opens the door to doing arbitrary manipulations
with scopes, including writing them into other scopes if necessary.

This is necessary for block scoping inside generators, since a yield in a
block scope may require writing out that scope to the frame, since it
would be live across the yield.

Reviewed By: avp

Differential Revision: D54934498

fbshipit-source-id: 2956525330a2f24465632fc134c0300a225b277e