Track start VariableScope in ResolveScopeInst
Summary:
In order to remove the requirement that all scope operands must be
`BaseScopeInst`, we need to explicitly encode the incoming scope to a
`ResolveScopeInst`, because we can no longer determine the incoming
`VariableScope` by casting the start scope operand to a `BaseScopeInst`.

We may eventually be able to remove this, if the metadata for the scope is
instead encoded in its type.

Reviewed By: avp

Differential Revision: D54932133

fbshipit-source-id: 3bb60f2f696fa2177f7e786346c67514b4006b5f