Define new JSGeneratorObject
Summary:
The only difference here compared to the existing JSGenerator is that
the inner function is simply a Callable. This means both HBC and native
code will be able to create a JSGeneratorObject. A new class is created
so we can use it incrementally, rather causing many compilation issues
in the existing generator codepath for the runtime.

I think adding 'Object' to the name also helps further clarify the
difference compared to a generator function.

Reviewed By: neildhar

Differential Revision: D53028337

fbshipit-source-id: 14ef217897995a8971b2ea15f62c0a6d18233719