Delete GeneratorInnerFunction in IR
Summary:
Move the distinction of GeneratorInnerFunction into the
DefinitionKind. This is a synthetic function that only we create,
so its definition can be treated as a separate kind.

Note: technically HoistStartGenerator and the check in
LowerArgumentsArray can be removed now. But, we are keeping this in
defensively, as the exact placing for LowerGenerators in the pipeline is
not guaranteed in perpetuity.

Reviewed By: avp

Differential Revision: D54420597

fbshipit-source-id: 03b06d3c2897ab9b183f3a3d100f21a495158cb5