Enforce scope usage using types
Summary:
Instead of trying to track the users of `BaseScopeInst` specifically,
check for any instruction that has an operand with Environment type.

Now that environments can be stored in stack locations or other
environments, this allows us to track unexpected usage even after they
are stored and loaded from these locations.

Reviewed By: avp

Differential Revision: D55947844

fbshipit-source-id: 8b8adb2e419aec4f03c207209d075eecc35c890d