Make TryEnd a terminator
Summary:
Change TryEnd into a terminator with two successors. The first successor
is the catch block, the second successor is a block to branch to.

The idea is that having the catch block as an explicit successor of
TryEnd allows us to perform (almost) correct liveness analysis inside
try blocks.

We don't know when an exception might occur in the try block, which
would "branch" to the catch block. However a try block can only be
terminated by a `Throw` or a `TryEnd`. By adding an explicit successor
to the latter, we encode the possibility that execution can continue in
the catch block.

(We will address `Throw` next).

Reviewed By: neildhar

Differential Revision: D56000872

fbshipit-source-id: a66dc0d2e62f34e49dfab6e4ff299029d4aa25d0