Optimise scopes during IRGen
Summary:
When emitting variable accesses during IRGen, try walking up the scope
chain to find a better match. This will avoid emitting ResolveScope for
accesses to the immediately enclosing scope, and in subsequent diffs,
will also ensure that accesses to scopes created in the same function
will not need ResolveScope.

Reviewed By: avp

Differential Revision: D56259269

fbshipit-source-id: 083016fe35d2558607d78b21a8c21a45a71b8d22