Fix IRVerifier labels
Summary:
There were 2 issues that led to labels being generated from IRVerifier
that were inconsistent with the dumped IR.

1. The temp namer being created didn't have its persistent state set.
This is an issue because we need to be able to produce labels from
arbitrary functions within the module from IRVerifier, so the
printer/namer must retain all this per-function information. Without
persistence, this information is lost upon entering a new function. This
is changed so that the temp namer is created with the same persistent
setting as the parameter passed in to the IRPrinter.
2. With persistence set, `newFunction` will call `nextGeneration` on the
current function state. This leads to unwanted interactions when we call
`newFunction` when producing an error message containing the BB label
and instruction label. This is because the first time we call it to get
the BB label, it increments the generation, and invalidates all the
instruction labels since none of them were visited- we only wanted to
get the BB label. So the next time that we call it to get the
instruction label, we get back an inconsistent number label. This issue
is solved by exposing a `restoreFunctionState` which will set the
correct function state, but not call `nextGeneration` on it.

Reviewed By: avp

Differential Revision: D55921933

fbshipit-source-id: ba540bd22fbafdd03dd2d6642a944df725389b9c