Maintain ordering of loaded scripts take 2
Summary:
Original Author: dsc@meta.com
Original Git: e03dbac7f72b353e69c519fe6a017384053f0338
Original Reviewed By: fbmal7
Original Revision: D50514612

Same change as D50236422, which had to be reverted due to it causing a
crash. This new diff fixes a previously unknown bug where CDPHandler
doesn't handle invalid SourceLocation properly.

We used to keep track of the loaded scripts in an `unordered_map`. This
means that when a client connects and sends us a Debugger.enable
message, the scriptParsed notifications that we send are not the same
order as when those loaded scripts appeared.

This is a problem because DevTools'
[DebuggerModel.ts](https://github.com/ChromeDevTools/devtools-frontend/blob/main/front_end/core/sdk/DebuggerModel.ts#L756) treats the
scriptParsed that appear later on as the newest version of the script.

For the React Native Fast Refresh scenario where the same script gets
modified and loaded multiple times, we want DevTools to treat the last
loaded version as the newest version.

Fixing this issue by switching to use a vector so that the behavior is
more like we buffered the loaded scripts and then send the scriptParsed
notifications in the order that they would have been sent if the
debugger client was already connected.

Reviewed By: mattbfb

Differential Revision: D56482793

fbshipit-source-id: 58fdf0e529055b8d1b938c4b4158c0fe0cbafd31