Introduce AsyncDebuggerAPI
Summary:
Original Author: dsc@meta.com
Original Git: 24dd5183cfb8956a7822340c659684d5d8283be4
Original Reviewed By: mattbfb
Original Revision: D51595285

Adding a new class AsyncDebuggerAPI that wraps the existing
DebuggerAPI. This class extracts some core concepts from the existing
CDPHandler so that we have a class that's not tied to CDP.

The main things that this class enables are:
1) Exposes an interrupt API and guarantees those interrupt callbacks
will be called exactly once.
2) Turns the didPause interaction into an asynchronous one. Consumers
of this class will register DebuggerEventCallback but isn't expected to
decide on a next debugger::Command while handling that callback.
3) Adds an API for unpausing the JavaScript program after a
DebuggerEventCallback has been invoked. Together with #2, this allows
someone to implement a debugging solution like CDP that's asynchronous
in nature.

The reason to introduce this class is so that we can contain most of
the multithread complexity in one place. Rather than having every
function in CDPHandler having to think about multithreading, CDPHandler
can be simplified.

Reviewed By: mattbfb

Differential Revision: D56482827

fbshipit-source-id: f51a77e36c5715df7541e3d13ac709b8a47052bb