Handle eval command & EvalComplete differently
Summary:
Original Author: dsc@meta.com
Original Git: 0618126c365bcd8db3db5d394680e34dc5d06eca
Original Reviewed By: mattbfb
Original Revision: D53257099

It does not make sense to broadcast the `EvalComplete` event to all
subscribers because the EvalComplete event is for a specific request.

Performing an eval as the next command from didPause is actually a bit
special. It's different because `lib/VM/Debugger/Debugger.cpp` will
remain in the debugger loop and any expression that gets evaluated will
not trigger recursive `didPause`. So the result of running an eval
command and getting the EvalComplete back can be seen as if those 2
things are just a synchronous function and the PauseReason has not
changed.

Reviewed By: mattbfb

Differential Revision: D56482849

fbshipit-source-id: 71acfb04d1ca33f04ca70679a70b25c02491f0d7