Add Map fast paths in Map constructor
Summary:
Reading from `Map` when the iterator is the default iterator can be made
faster by just iterating the entries directly.

Improves speed of a microbenchmark that just clones a
1000-element Map by ~8x.

NOTE: This can be further improved by bypassing the extra
hashing/expansion but that's going to require more effort. This is just
a first version of the fast path.

Reviewed By: fbmal7

Differential Revision: D56158304

fbshipit-source-id: 03b1de066ab06e5cbe91a8089095c4d93a8b9442