Make --label argument optional.
Summary:
The --label argument to bench-runner stores an extra string in the output json file, to go with the name of the execution engine.  This is displayed by bench-merge, so that if you compare results for hermes before/after a diff, you can get "hermes (before)" and "hermes (after)" after in the table bench-merge produces.

When I did this however, I mistakenly made the argument required.  It shouldn't be.

Reviewed By: avp

Differential Revision: D56526641

fbshipit-source-id: ad231e97f52590f40a128fedd1bc7db90f5a20fc