Error on redeclaring 'var' in nested scope
Summary:
`var` is hoisted to function scope, but it needs to be checked in its
position of declaration if it's in an inner scope to make sure that it
isn't redeclaring a let-like variable.

This also revealed that part of the `shadowed-var` test was invalid,
and it's actually supposed to error.

Reviewed By: neildhar

Differential Revision: D56643141

fbshipit-source-id: f830a39855eb078956fc6ec1d00411ad5ee1d51c