Make ObjectStackPromotion work with legacy objects
Summary:
Extend the logic of ObjectStackPromotion to handle legacy objects. The
core difference for handling legacy objects is about the lack of
structure. First, at this point in the pipeline, we don't immediately
know the layout of the object. The IR for objects is a beginning
AllocObject, followed by lots of StoreNewOwnInsts. So, in order to find
the layout of the object from this structure, we re-use existing code
from lowering that does exactly that.

Once we compute the object layout, we analyze all its users. If the only
users are loads and stores to *existing* properties, then the object
will be promoted to the stack.

Reviewed By: neildhar

Differential Revision: D56596944

fbshipit-source-id: 11227473de292a6678fbe8be02946c32a1a1e9b4