Compressed Doubles
Summary:
See related post introducing the idea:

https://fb.workplace.com/groups/271129191743522/permalink/663030872553350/

Synopsis: we've been saying "SMI", for "Small Integers", when we think
about the optimization we do to (mostly) get 32-bit HermesValues in
the heap.  But the test for whether a given double is a small integer
is fairly expensive.  This diff proposes a different perspective.
We're only using the "SMI" property to get a more compact
representation in the heap -- we don't, for example, maintain this,
and do special numerical operations for small integers.  (And I'm
grudgingly convinced of Tzvetan's view that the extra type tests would
offset any benefits this gets us.)

This view, that all we're trying to do is compress doubles, suggests a
different strategy.  Small integers, expressed as doubles, generally
have "short" mantissas -- some non-zero bits, then quite a lot of
trailing zeros.  The exponent and the high-order part of the mantissa
are in the first 32-bit word.  So we can just say a "compact double"
is one whose trailing 32 bits are zeros.  This is easy to test for,
and to convert to/from.  The set of numbers fitting this compressed
representation overlaps a lot with the previous small integers --
integers whose absolute values are < 2^17 fit.  But, it also covers a
bunch of non-integral doubles, depending on the exponent.  (We don't
know how useful this will be.)

I added a microbenchmark to emphasize these costs.

Reviewed By: neildhar

Differential Revision: D55044596

fbshipit-source-id: 1162ba2b65495107c6123d29e24c5e1187ca7571