Add fatal stubs for lazy compilation
Summary:
A first step toward re-enabling lazy compilation.
Delete some code that we're not planning to use.
Add some `hermes_fatal` stubs to indicate where lazy logic should be
inserted in the future.

`BytecodeGenerator` fatals when you try and compile a lazy function.
Of course, `-lazy` is currently a no-op so this actually does nothing
in this diff, it's just setup.

Reviewed By: dannysu

Differential Revision: D55655732

fbshipit-source-id: b753c9d29185d4cfd5d49737882e1ca5aa38029f