Add a function to reset the Module for lazy compilation
Summary:
To free memory and avoid regenerating bytecode for functions which have
already been compiled, add a function to "reset" the Module
after an iteration of lazy compilation of a function.
The function is called after BytecodeGenerator during lazy compilation,
and it removes generated functions, replaces the global function body
with `UnreachableInst`, and deletes any `VariableScope`s that are no
longer used.

Reviewed By: tmikov

Differential Revision: D55832354

fbshipit-source-id: 6a2c8d8cf7bf4b2b8be0586b96dccbfd28f9f177