Disable persistent flag on RuntimeModules in lazy compilation
Summary:
The underlying storage may be moved/reallocated under lazy compilation,
so the `RuntimeModule`s can't be marked as persistent.
We need to avoid storing lazy identifiers which point directly into the
`BytecodeModule`'s storage.

Allow the `Runtime` to query whether persistence is allowed on the given
`BCProviderFromSrc`.

Add a check in `runBytecode` to ensure the `persistent` flag is never
set on unsupported code.

Reviewed By: dannysu, tmikov

Differential Revision: D55702306

fbshipit-source-id: 58db82238df4474659455f535f21c8806b67f0eb