Append source info to DebugInfo fields and delete lexical info
Summary:
NOTE: This removes the lexical data from `DebugInfo`, but not the stored
offsets in `BytecodeFunction` (that's in the next diff). This avoids
breaking the build at all, even though the debugger won't run properly.

We don't need the lexical data in `DebugInfo` here any more - we'll
use the IR directly for `VariableScope` information much like for lazy
compilation.

So now all that's left is the source location information, which can be
appended to directly on a `StreamVector` provided we give some
additional APIs for updating the stored references as necessary.

Reviewed By: tmikov

Differential Revision: D55925672

fbshipit-source-id: fc3a30626dfe69938064b7bb97bd3f6039276b07