Begin lazy bytecode generation
Summary:
Add a `generateLazyFunction` to `BytecodeGenerator` which takes an
existing IR function and its ID and replaces the lazy stub with real
bytecode.

It will generate every new function in the IR module and then set any
references in the `BCProvider` that may have changed when the
`BytecodeModule` was mutated.
It will skip any lazy functions which have already had IDs assigned.

NOTE: This diff doesn't handle serialized literal instructions yet.
That's in a separate diff because it requires some other changes.

Reviewed By: tmikov

Differential Revision: D55702303

fbshipit-source-id: fa6573f2ac2bd3f1b792ced947287c6535e388e0