Use PointerValue for ObjectID (#1399) (#1399)
Summary:
Original Author: fumi@meta.com
Original Git: f62833355a0ddf2cc093d792483eb8efb0c5e45f
Original Reviewed By: neildhar
Original Revision: D57133457

Pull Request resolved: https://github.com/facebook/hermes/pull/1399

Use `PointerValue*` as a source of `SynthTrace::ObjectID` instead of
`UniqueID` from `Runtime` which is based on GC's unique ID system. This
will reflect better about the life time of each `jsi::Pointer` in the
record so that we can emulate the closer behavior to the original when
we replay the record.

Note that, the non-deterministic hooks for WeakRef needs to be changed
because the existing logic was relying on the fact that the ObjectID
for the referenced JSObject won't change between WeakRef creation and
deref time, which is no longer true.

Reviewed By: avp

Differential Revision: D57625936

fbshipit-source-id: 3bb20974731358e07e15e10b3e9fcc8aabaf4304