HBC instruction for uint8_t prop for GetByValue.
Summary:
I made measurements of frequency of bytecode execution in the
draw.react-native-forget.js benchmark.  GetByVal was 1.26% of executed
instructions.  Later, I measured how often some of the frequent
instructions had int literal operands in relevant positions.

This diff is the implementation of the optimization that suggested for
GetByVal.  When the property being accessed is a numeric literal
representable as a uint8_t, we do not generate a LoadLiteral
instruction for that.  Instead, we alter instruction selection to
generate a new variant instruction, GetByIndex, which has the uint8_t
literal as an immediate operand of the instruction.  This saves a
bytecode execution, translation to/from double, and a test on the
value of the int.

Perf:

I added a new microbenchmark, emphasizing this behavior, to the
benchmark suite: arrayReadByIndex.js.

Results for that:
```
BEFORE:  796, 798, 802
AFTER:   487, 484, 494
```
If we consider the min-time runs, that's an improvement of about 39%.

For draw.react-native-forget.js, the run-to-run variation is pretty
high, and the optimized bytecode is only 1.26% of bytecode instruction
executions.  But here are 5 runs:
```
BEFORE: 1436, 1420, 1402, 1422, 1407
AFTER:  1408, 1421, 1422, 1396, 1381
```

If we take averages, or min times, that's something like a 1%
decrease.  So that seems at least suggestive of a downwards trend in
times.

Reviewed By: tmikov

Differential Revision: D56910391

fbshipit-source-id: 07c616cbc98b82d9373a2191cb98574a5e798402