Fix building errors for hermesAppleMac
Summary:
We need to explicitly add the "AppleMac" suffix to build in arm64 mode
locally, currently it gives a few compiler errors: shadowing errors in
multiple LLVM headers (-Wshadow) and global constructor errors in
CompilerDriver (-Wglobal-constructors). Suppress them in the `hermes`
buck target (we suppressed them as well in Hermes).

Reviewed By: neildhar

Differential Revision: D57764717

fbshipit-source-id: a7dff46bded8e0f3f4d85e0c8384c77070f02609