Fix ASan SIOF false positive in llvh CommandLine
Summary:
ASan currently reports an SIOF here because we are iterating over and
accessing other globals during initialisation. However, this is known
to be safe, since the only globals being accessed are by definition the
ones that are known to already be initialised.

To work around this, disable the offending asserts when building with
ASan.

Reviewed By: tmikov

Differential Revision: D57888753

fbshipit-source-id: 02c6c100ff41f56043b8cf4909431ea7cafc978a