Don't set inline directives in SemanticResolver
Summary:
These directives should not have been set from
SemanticResolver to begin with. However, we decided it's
best to not force inlining for implicit field initializers
and implicit constructors in general. If there are many
fields in a typed class, that could be harmful. Instead, we
should let the optimization pass decide if lowering should
be done.

Reviewed By: tmikov

Differential Revision: D58441695

fbshipit-source-id: 35842748b02d484df86afe7364b89ef8f3e8a389