Store SHUnit pointer in NativeJSFunction
Summary:
Store a pointer to the SHUnit in NativeJSFunction. This allows us to
remove it from the function info, making it immutable and unit
independent, so we don't have to copy it into each copy of an SHUnit.

This does increase the size of functions, and there are a few ways we
could address that in the future:
1. Once we control the layout of emitted code, we could store the info
at some known offset right before the function pointer. This would allow
us to eliminate the function info pointer.
2. Instead of storing the unit and info pointers, store their indices.
For units, this is straightforward to do once we assign them indices,
but for function info, we will need to expose the info array through the
unit. Note that doing so is at odds with the first option, which
eliminates that array altogether.
3. We could do a separate GC heap allocation for the info and unit
pointers. If the function is instantiated many times, each copy can
reuse this allocation and point to it with a compressed pointer,
reducing the overall cost.

Reviewed By: tmikov

Differential Revision: D58649566

fbshipit-source-id: 743c4477cc139b6e71ce6dc9f5ec134bdc25c0ab