Add "description" to Flow schema generation
Summary:
Original Author: evanyeung@meta.com
Original Git: b60c98e271cbe10a97c31d5d67874ee970ccdb09
Original Reviewed By: tyao1
Original Revision: D58156907

@public
This diff adds description parsing to Flow based schema generation. The `description` field already existed in the docblock ir and we were just not setting it. This diff validates the docblock description then adds it to the field.

I also discovered that the location on the comment text is a relative location to the comment (so starts at 0) and always errors at the beginning of the file. I added another field to the comment attachment visitor to track the actual location of the comment.

Reviewed By: avp

Differential Revision: D58849741

fbshipit-source-id: 206f84b3615bc52222e022d9ffefcec34395fd8b