Move for loop body generation after control clauses
Summary:
The for loop body is considered a different statement for stepping
purposes than the control clauses.

This just moves the body generation to after the control clauses to
avoid interleaving and overlapping the statement index.

The change simply maintains behavior between Hermes and SH;
it's possible we'll want to revisit this (maybe each control clause
should be its own statement index?).

Reviewed By: tmikov

Differential Revision: D57356129

fbshipit-source-id: 490784de8317baaf4c413fcf87a44a8c17e18a79