Add EvalCompilationDataInst
Summary:
This instruction holds some similar data to LazyCompilationDataInst,
but it is kept around in functions which have already been compiled.
It also contains less data than `LazyCompilationDataInst` because it
doesn't have to point to existing source.

It will also be kept at the start of functions which we may want to
run local eval inside (via the debugger for now but eventually also with
`eval()`).

Reviewed By: neildhar

Differential Revision: D57356131

fbshipit-source-id: 44c667daf737f3aee533c00ac98a566f68d99a3d