Check for 'await' as an identifier in async arrow functions
Summary:
`await` is supposed to be an `AwaitExpression` in the parameters of an
async arrow function, including in the parameters of a normal arrow
function in the initializer of an async arrow function.

Due to having to reparse arrow parameters to detect async
arrow functions, we'd have to rescan the whole AST to error on this in
the parser, so just do it in SemanticResolver during
`resolveIdentifier`.

Reviewed By: fbmal7

Differential Revision: D58734791

fbshipit-source-id: d691f8e0ddb3f90b170e6802a1a9660588e8b76a