Add console bindings for shermes
Summary:
Add the ability to implement our own bindings implemented on top of JSI
in scripts running in shermes. This is necessary to get test262 running
against shermes again.

For now we build the console bindings both statically and dynamically,
and link it the same way we link hermesvm (as opposed to always
linking statically). This is simpler because we don't need to support
statically linking a C++ library across platforms, which we currently
only support on macOS for hermesvm.

Reviewed By: lavenzg

Differential Revision: D58002565

fbshipit-source-id: df1959ec5a8f5db2e5130271270807424e72e2b4