Specialize IsHermesValueConvertible to HermesValue
Summary:
The default implementation of `IsHermesValueConvertible` uses
`st::is_base_of` in order to check if type B is a subclass of type A,
where both are GCCell-inheriting classes. However, the issue is that
`is_base_of` needs the complete type definition in order to perform this
check. But, there are instances where it's not possible to see the full
type definition of a VM type, for instance in header files.
`IsHermesValueConvertible` is used, among other things, to enable to
implicit constructors for a Handle. The logic is that

The specific, motivating example for this diff is `Runtime.h` not being
able to see the full `JSObject`. It would be a mutual include problem
for `Runtime.h` to include `JSObject.h`, so it can't see its full type.
This means that you cannot use the implicit constructor for making a
`Handle<HermesValue>` from a `JSObject`, since the `is_base_of` check
will fail to compile. But it fails to compile for the "wrong" reason.
There's no easy way to fix this issue for `GCCell` types, aside from
using `vmcast`. But, when it comes to `HermesValue`, we can circumvent
this by not using `is_base_of` at all. We know that you can always
convert *to* a `HermesValue`. So we specialize on this pattern, so that
`IsHermesValueConvertible` can be more useful in header files and other
places that suffer from this same incomplete type issue.

Reviewed By: avp

Differential Revision: D59350272

fbshipit-source-id: 6952cabde320a7380100f13f7c57a81c285b8dd7