Add a 'noinline' directive -- the opposite of 'inline'.
Summary:
In writing a test in another diff, there was a situation where it
seemed like it would be convenient to have a 'noinline' directive.  (I
wanted to show that a non-inlined cross-module call found the callee
more efficiently than accessing an export object property.  I didn't
want to have to make the callee "complicated" to avoid inlining.)

This diff adds the directive.

Reviewed By: avp

Differential Revision: D58436718

fbshipit-source-id: da433c3bacaaf48a3719d50cf9dc6a8bb8b7b8bf