Add DenseMap::at, MapVector::at.
Summary:
The llvh version of llvm::DenseMap does not have a .at(key) operation.
(Treats the map as const; asserts that the key is present; returns a
const& reference to the value.). More recent versions of LLVM do:

https://llvm.org/doxygen/DenseMap_8h_source.html#l00211.

This diff proposes to copy that implementation into our version.

The stdlib versions of map::at have const and non-const versions;
there were use cases later in the stack where that the non-const version
was really useful, so added that as well.  Factored out duplicated
bodies using a macro.

Found I needed the same for MapVector, so added it there, too.

Reviewed By: avp

Differential Revision: D58549108

fbshipit-source-id: 33ffb458de7fac8a260176ff567734249b58b99c